Move problem details logic to dedicated component
