+++ frontend/src/app/annotate/annotation-input/annotation-input.component.html
@@ -8,9 +8,7 @@
 @if (form$ | async; as inputForm) {
 <div class="card">
     <div class="card-body">
-        @if (problemDetails$ | async; as details) {
-        <la-problem-details [problemDetails]="details" />
-        }
+        <la-problem-details />
         <div class="input-form-container">
             <la-premises-form class="subform" [form]="inputForm" />
             <la-knowledge-base-form class="subform" [form]="inputForm" />
