+++ frontend/src/app/annotate/annotation-input/annotation-input.component.ts
@@ -63,52 +63,7 @@ export class AnnotationInputComponent {
         initialValue: null,
     });
 
-    public problemDetails$ = this.problem$.pipe(
-        map((response) => this.extractDetails(response))
-    );
-
     public submit$ = new Subject<void>();
-
-    private getJudgement(response: ProblemResponse): Judgement {
-        // This should never happen, as we check for a problem in the calling
-        // function, but TypeScript does not know this.
-        if (!response.problem) {
-            return Judgement.UNKNOWN;
-        }
-        switch (response.type) {
-            case Dataset.SICK:
-                switch (response.problem.entailmentLabel) {
-                    case "ENTAILMENT":
-                        return Judgement.ENTAILMENT;
-                    case "CONTRADICTION":
-                        return Judgement.CONTRADICTION;
-                    case "NEUTRAL":
-                        return Judgement.NEUTRAL;
-                }
-            case Dataset.FRACAS:
-                switch (response.problem.fracasAnswer) {
-                    case "yes":
-                        return Judgement.ENTAILMENT;
-                    case "no":
-                        return Judgement.CONTRADICTION;
-                    case "unknown":
-                        return Judgement.NEUTRAL;
-                    case "undefined":
-                        return Judgement.UNKNOWN;
-                }
-            case Dataset.SNLI:
-                switch (response.problem.goldLabel) {
-                    case "entailment":
-                        return Judgement.ENTAILMENT;
-                    case "contradiction":
-                        return Judgement.CONTRADICTION;
-                    case "neutral":
-                        return Judgement.NEUTRAL;
-                    case "none":
-                        return Judgement.UNKNOWN;
-                }
-        }
-    }
 
     public faCheck = faCheck;
 
@@ -178,43 +133,5 @@ export class AnnotationInputComponent {
             }),
             kbItems: new FormArray<KnowledgeBaseItemsForm>([]),
         });
-    }
-
-    private extractDetails(
-        response: ProblemResponse | null
-    ): ProblemDetails | null {
-        if (!response?.problem) {
-            return null;
-        }
-        const judgement = this.getJudgement(response);
-        switch (response.type) {
-            case Dataset.SICK:
-                return {
-                    problemId: response.problem.pairId.toString(),
-                    dataset: response.type,
-                    judgement,
-                    section: null,
-                    subsection: null,
-                    comment: null,
-                };
-            case Dataset.FRACAS:
-                return {
-                    problemId: response.problem.fracasId.toString(),
-                    dataset: response.type,
-                    judgement,
-                    section: response.problem.sectionName,
-                    subsection: response.problem.subsectionName,
-                    comment: response.problem.note || null,
-                };
-            case Dataset.SNLI:
-                return {
-                    problemId: response.problem.pairId.toString(),
-                    dataset: response.type,
-                    judgement,
-                    section: null,
-                    subsection: null,
-                    comment: null,
-                };
-        }
     }
 }
